Spring Secure

+++ Spring_Secure/src/main/java/kr/co/hucloud/security/code/example/board/web/BoardController.java
@@ -6,7 +6,6 @@ import java.io.UnsupportedEncodingException;
 import java.util.List;
 import java.util.UUID;
 
-import javax.management.RuntimeErrorException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
@@ -26,7 +25,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.servlet.HttpServletBean;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.nhncorp.lucy.security.xss.XssFilter;
@@ -38,7 +36,7 @@ import io.github.seccoding.web.mimetype.ExtensionFilterFactory;
 @Controller
 public class BoardController {
 	
-	private static final String FILE_PATH = "D:/uploadFiles";
+	private static final String FILE_PATH = "E:/Source/uploadFiles";
 	
 	private BoardService boardService;
 	private ReplyService replyService;
@@ -59,48 +57,44 @@ public class BoardController {
 	@RequestMapping("/board/writeArticle")
 	public ModelAndView writeArticle(BoardVO board, HttpSession session) {
 		
-		// CSRF를 막기 위한 Token 비교
 		String sessionToken = (String) session.getAttribute(Session.CSRF_TOKEN);
-		if( !board.getToken().equals(sessionToken) ) {
+		if ( !board.getToken().equals(sessionToken) ) {
 			throw new RuntimeException("잘못된 접근입니다.");
 		}
-		
 		
 		String userId = ((MemberVO) session.getAttribute(Session.MEMBER)).getId();
 		board.setUserId(userId);
 		
 		MultipartFile uploadedFile = board.getFile();
 		
-		if( !uploadedFile.isEmpty() && uploadedFile != null ) {
+		if( uploadedFile != null && !uploadedFile.isEmpty() ) {
 			// 외부 URL 접근이 가능함.
 			
 			String fileName = UUID.randomUUID().toString();
 			String originFileName = uploadedFile.getOriginalFilename();
 			
-			File file = new File(FILE_PATH, fileName );
+			File file = new File(FILE_PATH, fileName);
 			try {
 				uploadedFile.transferTo(file);
 				board.setFileName(fileName);
 				board.setOriginFileName(originFileName);
 			} catch (IllegalStateException | IOException e) {
 				throw new RuntimeException(e.getMessage(), e);
-			}
-			finally {
-				if( file.exists() ) {
+			} finally {
+				if ( file.exists() ) {
 					
-					ExtensionFilter filter = 
-							ExtensionFilterFactory.getFilter(ExtFilter.JMIME_MAGIC);
+					ExtensionFilter filter = ExtensionFilterFactory.getFilter(ExtFilter.JMIME_MAGIC);
 					boolean isImageFile = filter.doFilter(
-							file.getAbsolutePath(), "image/bmp"
-												  , "image/png"
-												  , "image/jpeg"
-												  , "image/gif");
+							file.getAbsolutePath()
+							, "image/bmp"
+							, "image/png"
+							, "image/jpeg"
+							, "image/gif");
 					
-					if( !isImageFile ) {
+					if ( !isImageFile ) { 
 						file.delete();
 						board.setFileName("");
-						board.setOriginFileName(""); // 업로드를 잘못하면 지워버리기
-						boardService.modifyArticleInBoard(board);
+						board.setOriginFileName("");
 					}
 					
 				}
@@ -108,7 +102,7 @@ public class BoardController {
 		}
 		
 		// XSS 게시판의 제목과 내용에 대해 XSS HTML 인코딩
-		XssFilter filter = XssFilter.getInstance("lucy-xss-filter");
+		XssFilter filter = XssFilter.getInstance("lucy-xss-superset.xml");
 		board.setSubject( filter.doFilter( board.getSubject() ));
 		board.setContent( filter.doFilter( board.getContent() ));
 		
@@ -142,7 +136,7 @@ public class BoardController {
 		
 		if(boardList != null) {
 			// XSS 게시판의 제목과 내용에 대해 XSS HTML 인코딩
-			XssFilter filter = XssFilter.getInstance("lucy-xss-filter");
+			XssFilter filter = XssFilter.getInstance("lucy-xss-superset.xml");
 			for( BoardVO boardVO : boardList.getList() ) {
 				boardVO.setSubject( filter.doFilter( boardVO.getSubject() ));
 				boardVO.setContent( filter.doFilter( boardVO.getContent() ));
@@ -172,7 +166,7 @@ public class BoardController {
 		if(boardList != null) {
 			
 			// XSS 게시판의 제목과 내용에 대해 XSS HTML 인코딩
-			XssFilter filter = XssFilter.getInstance("lucy-xss-filter");
+			XssFilter filter = XssFilter.getInstance("lucy-xss-superset.xml");
 			for( BoardVO boardVO : boardList.getList() ) {
 				boardVO.setSubject( filter.doFilter( boardVO.getSubject() ));
 				boardVO.setContent( filter.doFilter( boardVO.getContent() ));
@@ -192,38 +186,62 @@ public class BoardController {
 	}
 	
 	@RequestMapping("/board/recommend/{id}")
-	public ModelAndView recommend(@PathVariable String id
-								, @RequestParam String token
-								, HttpSession session) {
+	public ModelAndView recommend(
+			@PathVariable String id
+			, @RequestParam String token
+			, HttpSession session) {
 		
-		// CSRF를 막기위한 Token 비교
 		String sessionToken = (String) session.getAttribute(Session.CSRF_TOKEN);
-		if( !sessionToken.equals(token) ) {
+		if ( !sessionToken.equals(token) ) {
 			throw new RuntimeException("잘못된 접근입니다.");
 		}
-		
-		
 		boardService.updateRecommend(id);
 		
 		return new ModelAndView("redirect:/board/article/" + id);
 	}
 	
 	@RequestMapping("/board/delete/{id}")
-	public ModelAndView delete(@PathVariable String id) {
-		//FIXME 잘못된 접근 제어 --> 요청자의 ID와 글 작성자의 ID를 비교해야 함
-		replyService.deleteReply(id);
-		boardService.deleteArticle(id);
+	public ModelAndView delete(
+			@PathVariable String id
+			, HttpSession session) {
+		// 잘못된 접근 제어 --> 요청자의 ID와 글 작성자의 ID를 비교해야 함
+		MemberVO memberVO = (MemberVO) session.getAttribute(Session.MEMBER);
+		String sessionMemberId = memberVO.getId();
+		
+		// 글 작성자 ID
+		BoardVO boardVO = boardService.getBoardById(id).boardAt(0);
+		String boardMemberId = boardVO.getUserId();
+		
+		if ( sessionMemberId.equals(boardMemberId) ) {
+			replyService.deleteReply(id);
+			boardService.deleteArticle(id);
+		}
+		
 		return new ModelAndView("redirect:/board/list");
 	}
 	
 	@RequestMapping("/board/modify/{id}")
-	public ModelAndView modify(@PathVariable String id) {
-		//FIXME 잘못된 접근 제어 --> 요청자의 ID와 글 작성자의 ID를 비교해야 함
-		BoardListVO boardList = boardService.getBoardById(id);
+	public ModelAndView modify(
+			@PathVariable String id
+			, HttpSession session) {
+		// 잘못된 접근 제어 --> 요청자의 ID와 글 작성자의 ID를 비교해야 함
+		
+		// 잘못된 접근 제어 --> 요청자의 ID와 글 작성자의 ID를 비교해야 함
+		MemberVO memberVO = (MemberVO) session.getAttribute(Session.MEMBER);
+		String sessionMemberId = memberVO.getId();
+				
+
+		BoardVO boardVO = boardService.getBoardById(id).boardAt(0);
+		String boardMemberId = boardVO.getUserId();
+		
+		if ( !sessionMemberId.equals(boardMemberId) ) {
+			return new ModelAndView("redirect:/board/list");
+		}
+		
 		ModelAndView view = new ModelAndView("board/modify");
 		
-		if(boardList != null) {
-			view.addObject("article", (BoardVO)boardList.getList().get(0));
+		if(boardVO != null) {
+			view.addObject("article", boardVO);
 		}
 		
 		return view;
@@ -235,6 +253,13 @@ public class BoardController {
 		String userId = ((MemberVO) session.getAttribute(Session.MEMBER)).getId();
 		board.setUserId(userId);
 		
+		BoardVO boardVO = boardService.getBoardById(board.getId() + "").boardAt(0);
+		String boardMemberId = boardVO.getUserId();
+		
+		if ( !userId.equals(boardMemberId) ) {
+			return new ModelAndView("redirect:/board/list");
+		}
+		
 		MultipartFile uploadedFile = board.getFile();
 		
 		if(board.isDelete()) {
@@ -244,36 +269,34 @@ public class BoardController {
 			board.setOriginFileName("");
 		}
 		
-		if( !uploadedFile.isEmpty() && uploadedFile != null ) {
+		if( uploadedFile != null && !uploadedFile.isEmpty() ) {
 			// 외부 URL 접근이 가능함.
 			
 			String fileName = UUID.randomUUID().toString();
 			String originFileName = uploadedFile.getOriginalFilename();
 			
-			File file = new File(FILE_PATH, fileName );
+			File file = new File(FILE_PATH, fileName);
 			try {
 				uploadedFile.transferTo(file);
 				board.setFileName(fileName);
 				board.setOriginFileName(originFileName);
 			} catch (IllegalStateException | IOException e) {
 				throw new RuntimeException(e.getMessage(), e);
-			}
-			finally {
-				if( file.exists() ) {
-					// Type 체크
-					ExtensionFilter filter = 
-							ExtensionFilterFactory.getFilter(ExtFilter.JMIME_MAGIC);
-					boolean isImageFile = filter.doFilter(
-							file.getAbsolutePath(), "image/bmp"
-												  , "image/png"
-												  , "image/jpeg"
-												  , "image/gif");
+			} finally {
+				if ( file.exists() ) {
 					
-					if( !isImageFile ) {
-						file.delete(); // 타입이 안맞으면 지워버린다.
+					ExtensionFilter filter = ExtensionFilterFactory.getFilter(ExtFilter.JMIME_MAGIC);
+					boolean isImageFile = filter.doFilter(
+							file.getAbsolutePath()
+							, "image/bmp"
+							, "image/png"
+							, "image/jpeg"
+							, "image/gif");
+					
+					if ( !isImageFile ) { 
+						file.delete();
 						board.setFileName("");
-						board.setOriginFileName(""); // 업로드를 잘못하면 지워버리기
-						boardService.modifyArticleInBoard(board);
+						board.setOriginFileName("");
 					}
 					
 				}
@@ -287,21 +310,20 @@ public class BoardController {
 	}
 	
 	@RequestMapping("/download/{boardId}")
-	public void download( @PathVariable String boardId
-							, HttpServletRequest request
-							, HttpServletResponse response) {
+	public void download(
+			@PathVariable String boardId
+			, HttpServletRequest request
+			, HttpServletResponse response) {
 		
-		BoardVO boardVO = boardService.getBoardById(boardId).boardAt(0);
-		String fileName = boardVO.getFileName();
-		String originFileName = boardVO.getOriginFileName();
+		BoardVO board = boardService.getBoardById(boardId).boardAt(0);
+		String fileName = board.getFileName();
+		String originFileName = board.getOriginFileName();
 		
 		try {
-			new DownloadUtil(FILE_PATH + File.separator + fileName)
-						.download(request, response, originFileName);
+			new DownloadUtil(FILE_PATH + File.separator + fileName).download(request, response, originFileName);
 		} catch (UnsupportedEncodingException e) {
 			throw new RuntimeException(e.getMessage(), e);
 		}
-		
 	}
 	
 }
